implement config.toml inside gql
