+++ core/src/graphman/shard.rs
@@ -1,0 +1,124 @@
+use std::collections::BTreeMap;
+
+use anyhow::anyhow;
+use serde::{Deserialize, Serialize};
+
+use graph::prelude::anyhow::{self, Result};
+use url::Url;
+
+use super::config::{Opt, PoolSize, Replica};
+use anyhow::Context;
+use graph_store_postgres::{DeploymentPlacer, Shard as ShardName, PRIMARY_SHARD};
+
+fn one() -> usize {
+    1
+}
+
+pub fn validate_name(s: &str) -> Result<()> {
+    if s.is_empty() {
+        return Err(anyhow!("names must not be empty"));
+    }
+    if s.len() > 30 {
+        return Err(anyhow!(
+            "names can be at most 30 characters, but `{}` has {} characters",
+            s,
+            s.len()
+        ));
+    }
+
+    if !s
+        .chars()
+        .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '-')
+    {
+        return Err(anyhow!(
+            "name `{}` is invalid: names can only contain lowercase alphanumeric characters or '-'",
+            s
+        ));
+    }
+    Ok(())
+}
+
+/// Replace the host portion of `url` and return a new URL with `host`
+/// as the host portion
+///
+/// Panics if `url` is not a valid URL (which won't happen in our case since
+/// we would have paniced before getting here as `url` is the connection for
+/// the primary Postgres instance)
+pub fn replace_host(url: &str, host: &str) -> String {
+    let mut url = match Url::parse(url) {
+        Ok(url) => url,
+        Err(_) => panic!("Invalid Postgres URL {}", url),
+    };
+    if let Err(e) = url.set_host(Some(host)) {
+        panic!("Invalid Postgres url {}: {}", url, e);
+    }
+    String::from(url)
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct Shard {
+    pub connection: String,
+    #[serde(default = "one")]
+    pub weight: usize,
+    #[serde(default)]
+    pub pool_size: PoolSize,
+    #[serde(default = "PoolSize::five")]
+    pub fdw_pool_size: PoolSize,
+    #[serde(default)]
+    pub replicas: BTreeMap<String, Replica>,
+}
+
+impl Shard {
+    pub fn validate(&mut self, name: &str) -> Result<()> {
+        ShardName::new(name.to_string()).map_err(|e| anyhow!(e))?;
+
+        self.connection = shellexpand::env(&self.connection)?.into_owned();
+
+        if matches!(self.pool_size, PoolSize::None) {
+            return Err(anyhow!("missing pool size definition for shard `{}`", name));
+        }
+
+        self.pool_size
+            .validate(name == PRIMARY_SHARD.as_str(), &self.connection)?;
+        for (name, replica) in self.replicas.iter_mut() {
+            validate_name(name).context("illegal replica name")?;
+            replica.validate(name == PRIMARY_SHARD.as_str(), &self.pool_size)?;
+        }
+
+        let no_weight =
+            self.weight == 0 && self.replicas.values().all(|replica| replica.weight == 0);
+        if no_weight {
+            return Err(anyhow!(
+                "all weights for shard `{}` are 0; \
+                remove explicit weights or set at least one of them to a value bigger than 0",
+                name
+            ));
+        }
+        Ok(())
+    }
+
+    pub fn from_opt(is_primary: bool, opt: &Opt) -> Result<Self> {
+        let postgres_url = opt
+            .postgres_url
+            .as_ref()
+            .expect("validation checked that postgres_url is set");
+        let pool_size = PoolSize::Fixed(opt.store_connection_pool_size);
+        pool_size.validate(is_primary, postgres_url)?;
+        let mut replicas = BTreeMap::new();
+        for (i, host) in opt.postgres_secondary_hosts.iter().enumerate() {
+            let replica = Replica {
+                connection: replace_host(postgres_url, host),
+                weight: opt.postgres_host_weights.get(i + 1).cloned().unwrap_or(1),
+                pool_size: pool_size.clone(),
+            };
+            replicas.insert(format!("replica{}", i + 1), replica);
+        }
+        Ok(Self {
+            connection: postgres_url.clone(),
+            weight: opt.postgres_host_weights.first().cloned().unwrap_or(1),
+            pool_size,
+            fdw_pool_size: PoolSize::five(),
+            replicas,
+        })
+    }
+}
