+++ core/src/lib.rs
@@ -12,3 +12,6 @@ pub use crate::subgraph::{
 
 #[macro_use]
 extern crate diesel;
+
+#[macro_use]
+extern crate juniper;
