+++ graph/src/components/store/mod.rs
@@ -915,7 +915,9 @@ pub struct StoredDynamicDataSource {
 /// identifier only has meaning in the context of a specific instance of
 /// graph-node. Only store code should ever construct or consume it; all
 /// other code passes it around as an opaque token.
-#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
+#[derive(
+    Copy, Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Hash, juniper::GraphQLScalarValue,
+)]
 pub struct DeploymentId(pub i32);
 
 impl Display for DeploymentId {
