+++ graph/src/data/subgraph/mod.rs
@@ -12,6 +12,7 @@ pub use features::{SubgraphFeature, SubgraphFeatureValidationError};
 
 use anyhow::{anyhow, Context, Error};
 use futures03::{future::try_join3, stream::FuturesOrdered, TryStreamExt as _};
+use juniper::{GraphQLObject, GraphQLScalarValue};
 use semver::Version;
 use serde::{de, ser};
 use serde_yaml;
@@ -71,7 +72,7 @@ where
 
 /// The IPFS hash used to identifiy a deployment externally, i.e., the
 /// `Qm..` string that `graph-cli` prints when deploying to a subgraph
-#[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Default)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Default, GraphQLScalarValue)]
 pub struct DeploymentHash(String);
 
 impl stable_hash_legacy::StableHash for DeploymentHash {
