+++ graph/src/data/subgraph/schema.rs
@@ -2,6 +2,7 @@
 
 use anyhow::{anyhow, bail, Error};
 use hex;
+use juniper::GraphQLObject;
 use lazy_static::lazy_static;
 use rand::rngs::OsRng;
 use rand::Rng;
@@ -22,7 +23,7 @@ lazy_static! {
     pub static ref POI_OBJECT: EntityType = EntityType::new("Poi$".to_string());
 }
 
-#[derive(Copy, Clone, PartialEq, Eq, Debug, Deserialize)]
+#[derive(Copy, Clone, PartialEq, Eq, Debug, Deserialize, juniper::GraphQLEnum)]
 #[serde(rename_all = "lowercase")]
 pub enum SubgraphHealth {
     /// Syncing without errors.
@@ -241,7 +242,7 @@ impl SubgraphManifestEntity {
     }
 }
 
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
+#[derive(Clone, Debug, Default, PartialEq, Eq, GraphQLObject)]
 pub struct SubgraphError {
     pub subgraph_id: DeploymentHash,
     pub message: String,
