+++ node/src/manager/commands/config.rs
@@ -13,7 +13,8 @@ use graph::{
 use graph_chain_ethereum::{NodeCapabilities, ProviderEthRpcMetrics};
 use graph_store_postgres::DeploymentPlacer;
 
-use crate::{chain::create_ethereum_networks_for_chain, config::Config};
+use graph_core::graphman::chain::create_ethereum_networks_for_chain;
+use graph_core::graphman::config::Config;
 
 pub fn place(placer: &dyn DeploymentPlacer, name: &str, network: &str) -> Result<(), Error> {
     match placer.place(name, network).map_err(|s| anyhow!(s))? {
